extract _cardKeys() to src/functions
