+++ src/functions/index.ts
@@ -1,2 +1,3 @@
+export * from './cardKeys';
 export * from './fisherYates';
 export * from './getRandomIndex';
